(feature) add config IsOnlyUpload

+++ main.go
@@ -27,11 +27,13 @@ func main() {
 	}
 	if !config.IsIncremental || incremental == nil {
 		// TODO: delete after upload
-		fmt.Println("---- [delete] ---->")
-		deleteErrs := operation.DeleteObjects(config.Bucket)
-		utils.LogErrors(deleteErrs)
-		fmt.Println("<---- [delete end] ----")
-		fmt.Println()
+		if !config.IsOnlyUpload {
+			fmt.Println("---- [delete] ---->")
+			deleteErrs := operation.DeleteObjects(config.Bucket)
+			utils.LogErrors(deleteErrs)
+			fmt.Println("<---- [delete end] ----")
+			fmt.Println()
+		}
 	}
 
 	records := utils.WalkDir(config.Folder)
@@ -40,14 +42,16 @@ func main() {
 	uploaded, uploadErrs := operation.UploadObjects(config.Folder, config.Bucket, records, incremental)
 	utils.LogErrors(uploadErrs)
 	fmt.Println("<---- [upload end] ----")
-	fmt.Println()
+	fmt.Println(config.IsOnlyUpload)
 
 	if config.IsIncremental && incremental != nil {
-		fmt.Println("---- [delete] ---->")
-		deleteErrs := operation.DeleteObjectsIncremental(config.Bucket, incremental)
-		utils.LogErrors(deleteErrs)
-		fmt.Println("<---- [delete end] ----")
-		fmt.Println()
+		if !config.IsOnlyUpload {
+			fmt.Println("---- [delete] ---->")
+			deleteErrs := operation.DeleteObjectsIncremental(config.Bucket, incremental)
+			utils.LogErrors(deleteErrs)
+			fmt.Println("<---- [delete end] ----")
+			fmt.Println()
+		}
 	}
 
 	if config.IsIncremental {
